refactor: modularize sidebar
