+++ pages/advanced.tsx
@@ -16,34 +16,49 @@ export default function Advanced(): JSX.Element {
   const { probeData, handleAddProbe } = useContext(ProbeContext);
   const [isModalVisible, setIsModalVisible] = useState<boolean>(false);
 
-  const renderProbe = () => {
-    return (
-      <div>
-        {probeData.map((probe: Probe) => (
-          <ProbeCard probe={probe} key={probe.id} id={probe.id} />
-        ))}
-        <button
-          onClick={handleAddProbe}
-          className="w-full border-4 border-dashed rounded-md p-4">
-          <p>Add another probe</p>
-        </button>
-      </div>
-    );
+  type Menu = 'Probe' | 'Notification';
+  const [activeMenu, setActiveMenu] = useState<Menu>('Probe');
+  const handleSetActiveMenu = (
+    event: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
+    menu: Menu
+  ) => {
+    event.preventDefault();
+
+    setActiveMenu(menu);
   };
 
-  const renderNotification = () => {
-    return (
-      <div>
-        {notificationData.map((notif: Notification) => (
-          <NotifCard key={notif.id} id={notif.id} type={notif.type} />
-        ))}
-        <button
-          onClick={handleAddNotification}
-          className="w-full border-4 border-dashed rounded-md p-4">
-          <p>Add another notification</p>
-        </button>
-      </div>
-    );
+  const render = () => {
+    if (activeMenu === 'Notification') {
+      return (
+        <div>
+          {notificationData.map((notif: Notification) => (
+            <NotifCard key={notif.id} id={notif.id} type={notif.type} />
+          ))}
+          <button
+            onClick={handleAddNotification}
+            className="w-full border-4 border-dashed rounded-md p-4">
+            <p>Add another notification</p>
+          </button>
+        </div>
+      );
+    }
+
+    if (activeMenu === 'Probe') {
+      return (
+        <div>
+          {probeData.map((probe: Probe) => (
+            <ProbeCard probe={probe} key={probe.id} id={probe.id} />
+          ))}
+          <button
+            onClick={handleAddProbe}
+            className="w-full border-4 border-dashed rounded-md p-4">
+            <p>Add another probe</p>
+          </button>
+        </div>
+      );
+    }
+
+    return <>Nothing to render</>;
   };
 
   return (
@@ -57,11 +72,15 @@ export default function Advanced(): JSX.Element {
           onClose={() => setIsModalVisible(false)}
         />
       </div>
-      <Sidebar>
-        {({ activeMenu }) =>
-          activeMenu === 'Probe' ? renderProbe() : renderNotification()
-        }
-      </Sidebar>
+      <div className="flex flex-col md:flex-row space-x-0 space-y-8 md:space-y-0 md:space-x-8">
+        <div className="w-full md:w-3/12">
+          <Sidebar
+            activeMenu={activeMenu}
+            onMenuChange={(event, menu) => handleSetActiveMenu(event, menu)}
+          />
+        </div>
+        <div className="w-full md:w-9/12">{render()}</div>
+      </div>
     </Layout>
   );
 }
